wip [ci skip]

+++ tools/shared.py
@@ -2871,6 +2871,7 @@ class Building(object):
                      '--dwarfdump=' + LLVM_DWARFDUMP,
                      '-o',  map_file]
     check_call(sourcemap_cmd)
+    #run_process([os.path.expanduser('~/Dev/wtmaps-utils/target/debug/wtmaps'), wasm_file, '-o', map_file])
 
   @staticmethod
   def get_binaryen_feature_flags():
@@ -2909,8 +2910,6 @@ class Building(object):
       # binaryen transformations. First, before running the command generate
       # an source map that is the identity transformation, i.e., represents
       # the locations of things before we change anything.
-      WTMAPS = os.path.expanduser('~/Dev/wtmaps-utils/target/debug/wtmaps')
-      WDWARF_CP = os.path.expanduser('~/Dev/wtmaps-utils/target/debug/wdwarf-cp')
       temp_files = configuration.get_temp_files()
       # make sure the input and output files are different, as we will need
       # them both around for the fixing up of the dwarf info
@@ -2921,7 +2920,6 @@ class Building(object):
       identity_map = temp_files.get('.map').name
       # TODO: we could detect if we already created a map for this wasm
       Building.emit_wasm_source_map(infile, identity_map)
-      #run_process([WTMAPS, infile, '-o', identity_map])
       output_map = outfile + '.map'
       cmd += ['--input-source-map=' + identity_map]
       cmd += ['--output-source-map=' + output_map]
@@ -2933,19 +2931,17 @@ class Building(object):
       run_process(cmd)
       ret = None
 
-    if preserve_dwarf:
+    if 0 and preserve_dwarf:
       # The wasm has been modified, and we emitted a new source map. Adjust
       # the debug info in the new wasm file.
       # We are emitting the maximum amount of debug info; in the wasm backend,
       # our wasm file contains DWARF sections. Keep them valid through
       # binaryen transformations
-      WTMAPS = os.path.expanduser('~/Dev/wtmaps-utils/target/debug/wtmaps')
       WDWARF_CP = os.path.expanduser('~/Dev/wtmaps-utils/target/debug/wdwarf-cp')
       temp_wasm = temp_files.get('.wasm').name
       run_process([WDWARF_CP, infile, '-o', temp_wasm, '-m', output_map, '-w', outfile])
       shutil.copyfile(temp_wasm, outfile)
       Building.emit_wasm_source_map(outfile, output_map)
-      #run_process([WTMAPS, outfile, '-o', output_map])
 
     return ret
